test: moving rust tests out of files
While it is good practice in Rust to include unit tests with the code files we've moved them out as integration tests instead. We will potentially put in some unit testing in Rust code files later, but overall this practice creates over-large code files for our AI analysis. Future improvements to our Rust plugin will address this issue in a more practical way like ignoring test sections within Rust files during analysis of the code itself.

Resolves #13
