+++ src/core/matrix.rs
@@ -62,7 +62,7 @@ pub struct CodeElement {
     pub metadata: serde_json::Value, // Plugin-specific data
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum ElementType {
     Function,
     Method,
@@ -85,7 +85,7 @@ pub struct Import {
     pub import_type: ImportType,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum ImportType {
     Standard,   // from standard library
     ThirdParty, // from external dependency
@@ -103,7 +103,7 @@ pub struct Relationship {
     pub strength: f32, // 0.0 to 1.0 - how strong is this relationship
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum RelationshipType {
     Import,
     Call,          // Function call across files
@@ -131,7 +131,7 @@ pub struct ExternalDependency {
     pub source_file: PathBuf,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum DependencyType {
     Runtime,
     Development,
@@ -422,273 +422,4 @@ pub struct ProjectMetrics {
     pub total_relationships: usize,
     pub highly_coupled_files: Vec<(PathBuf, usize)>,
     pub languages: Vec<String>,
-}
-
-// Unit tests
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::path::PathBuf;
-    use tempfile::TempDir;
-
-    // Helper function to create a test FileNode
-    fn create_test_file_node(path: &str, plugin: &str) -> FileNode {
-        FileNode {
-            path: PathBuf::from(path),
-            relative_path: PathBuf::from(path),
-            hash: "test_hash_123".to_string(),
-            size_bytes: 1024,
-            plugin: plugin.to_string(),
-            language: Some(plugin.to_string()),
-            is_text: true,
-            elements: vec![],
-            imports: vec![],
-            exports: vec![],
-            file_summary: Some("Test file summary".to_string()),
-        }
-    }
-
-    // Helper function to create a test Relationship
-    fn create_test_relationship(from: &str, to: &str) -> Relationship {
-        Relationship {
-            from_file: PathBuf::from(from),
-            to_file: PathBuf::from(to),
-            relationship_type: RelationshipType::Import,
-            details: "test import".to_string(),
-            line_number: Some(10),
-            strength: 0.8,
-        }
-    }
-
-    #[test]
-    fn test_project_matrix_creation() {
-        let project_root = PathBuf::from("/test/project");
-        let matrix = ProjectMatrix::new(project_root.clone());
-
-        assert_eq!(matrix.metadata.project_root, project_root);
-        assert_eq!(matrix.metadata.total_files, 0);
-        assert_eq!(matrix.metadata.total_size_bytes, 0);
-        assert!(matrix.files.is_empty());
-        assert!(matrix.relationships.is_empty());
-        assert!(matrix.external_dependencies.is_empty());
-        assert_eq!(matrix.metadata.csd_version, env!("CARGO_PKG_VERSION"));
-    }
-
-    #[test]
-    fn test_add_file() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-        let file_node = create_test_file_node("src/main.rs", "rust");
-
-        matrix.add_file(file_node.clone());
-
-        assert_eq!(matrix.metadata.total_files, 1);
-        assert_eq!(matrix.metadata.total_size_bytes, 1024);
-        assert!(matrix.metadata.plugins_used.contains(&"rust".to_string()));
-        assert!(matrix.files.contains_key(&PathBuf::from("src/main.rs")));
-    }
-
-    #[test]
-    fn test_add_multiple_files_different_plugins() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-        let rust_file = create_test_file_node("src/main.rs", "rust");
-        let python_file = create_test_file_node("script.py", "python");
-
-        matrix.add_file(rust_file);
-        matrix.add_file(python_file);
-
-        assert_eq!(matrix.metadata.total_files, 2);
-        assert_eq!(matrix.metadata.total_size_bytes, 2048);
-        assert_eq!(matrix.metadata.plugins_used.len(), 2);
-        assert!(matrix.metadata.plugins_used.contains(&"rust".to_string()));
-        assert!(matrix.metadata.plugins_used.contains(&"python".to_string()));
-    }
-
-    #[test]
-    fn test_add_relationship() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-        let relationship = create_test_relationship("src/main.rs", "src/lib.rs");
-
-        matrix.add_relationship(relationship.clone());
-
-        assert_eq!(matrix.relationships.len(), 1);
-        let added_rel = &matrix.relationships[0];
-        assert_eq!(added_rel.from_file, PathBuf::from("src/main.rs"));
-        assert_eq!(added_rel.to_file, PathBuf::from("src/lib.rs"));
-        assert_eq!(added_rel.strength, 0.8);
-    }
-
-    #[test]
-    fn test_add_external_dependency() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-        let dependency = ExternalDependency {
-            name: "serde".to_string(),
-            version: Some("1.0.0".to_string()),
-            ecosystem: "cargo".to_string(),
-            dependency_type: DependencyType::Runtime,
-            source_file: PathBuf::from("Cargo.toml"),
-        };
-
-        matrix.add_external_dependency(dependency.clone());
-
-        assert_eq!(matrix.external_dependencies.len(), 1);
-        let added_dep = &matrix.external_dependencies[0];
-        assert_eq!(added_dep.name, "serde");
-        assert_eq!(added_dep.version, Some("1.0.0".to_string()));
-        assert_eq!(added_dep.ecosystem, "cargo");
-    }
-
-    #[test]
-    fn test_get_files_by_plugin() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-        let rust_file1 = create_test_file_node("src/main.rs", "rust");
-        let rust_file2 = create_test_file_node("src/lib.rs", "rust");
-        let python_file = create_test_file_node("script.py", "python");
-
-        matrix.add_file(rust_file1);
-        matrix.add_file(rust_file2);
-        matrix.add_file(python_file);
-
-        let rust_files = matrix.get_files_by_plugin("rust");
-        let python_files = matrix.get_files_by_plugin("python");
-        let missing_files = matrix.get_files_by_plugin("javascript");
-
-        assert_eq!(rust_files.len(), 2);
-        assert_eq!(python_files.len(), 1);
-        assert_eq!(missing_files.len(), 0);
-    }
-
-    #[tokio::test]
-    async fn test_save_and_load_matrix() {
-        // Create a temporary directory for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let matrix_path = temp_dir.path().join("test_matrix.json");
-
-        // Create and populate a matrix
-        let mut original_matrix = ProjectMatrix::new(PathBuf::from("/test/project"));
-        let file_node = create_test_file_node("src/main.rs", "rust");
-        let relationship = create_test_relationship("src/main.rs", "src/lib.rs");
-
-        original_matrix.add_file(file_node);
-        original_matrix.add_relationship(relationship);
-
-        // Save the matrix
-        original_matrix
-            .save(&matrix_path)
-            .await
-            .expect("Failed to save matrix");
-
-        // Verify the file was created
-        assert!(matrix_path.exists());
-
-        // Load the matrix
-        let loaded_matrix = ProjectMatrix::load(&matrix_path)
-            .await
-            .expect("Failed to load matrix");
-
-        // Verify the loaded matrix matches the original
-        assert_eq!(
-            loaded_matrix.metadata.project_root,
-            original_matrix.metadata.project_root
-        );
-        assert_eq!(
-            loaded_matrix.metadata.total_files,
-            original_matrix.metadata.total_files
-        );
-        assert_eq!(loaded_matrix.files.len(), original_matrix.files.len());
-        assert_eq!(
-            loaded_matrix.relationships.len(),
-            original_matrix.relationships.len()
-        );
-
-        // Verify specific file exists
-        assert!(loaded_matrix
-            .files
-            .contains_key(&PathBuf::from("src/main.rs")));
-    }
-
-    #[test]
-    fn test_find_dependencies_and_dependents() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-
-        // Create files
-        let main_file = create_test_file_node("src/main.rs", "rust");
-        let lib_file = create_test_file_node("src/lib.rs", "rust");
-        let utils_file = create_test_file_node("src/utils.rs", "rust");
-
-        // Add files to matrix
-        matrix.add_file(main_file);
-        matrix.add_file(lib_file);
-        matrix.add_file(utils_file);
-
-        // Create relationships: main.rs -> lib.rs -> utils.rs
-        let rel1 = create_test_relationship("src/main.rs", "src/lib.rs");
-        let rel2 = create_test_relationship("src/lib.rs", "src/utils.rs");
-
-        matrix.add_relationship(rel1);
-        matrix.add_relationship(rel2);
-
-        // Test dependencies (what main.rs depends on)
-        let main_deps = matrix.find_dependencies(&PathBuf::from("src/main.rs"));
-        assert_eq!(main_deps.len(), 1);
-        assert_eq!(main_deps[0].path, PathBuf::from("src/lib.rs"));
-
-        // Test dependents (what depends on lib.rs)
-        let lib_dependents = matrix.find_dependents(&PathBuf::from("src/lib.rs"));
-        assert_eq!(lib_dependents.len(), 1);
-        assert_eq!(lib_dependents[0].path, PathBuf::from("src/main.rs"));
-
-        // Test file with no dependencies
-        let utils_deps = matrix.find_dependencies(&PathBuf::from("src/utils.rs"));
-        assert_eq!(utils_deps.len(), 0);
-
-        // Test file with no dependents
-        let utils_dependents = matrix.find_dependents(&PathBuf::from("src/utils.rs"));
-        assert_eq!(utils_dependents.len(), 1);
-        assert_eq!(utils_dependents[0].path, PathBuf::from("src/lib.rs"));
-    }
-
-    #[test]
-    fn test_calculate_metrics() {
-        let mut matrix = ProjectMatrix::new(PathBuf::from("/test"));
-
-        // Create a small project structure
-        let files = vec![
-            create_test_file_node("src/main.rs", "rust"),
-            create_test_file_node("src/lib.rs", "rust"),
-            create_test_file_node("src/utils.rs", "rust"),
-            create_test_file_node("script.py", "python"),
-        ];
-
-        for file in files {
-            matrix.add_file(file);
-        }
-
-        // Add some relationships to make lib.rs highly coupled
-        let relationships = vec![
-            create_test_relationship("src/main.rs", "src/lib.rs"),
-            create_test_relationship("script.py", "src/lib.rs"),
-            create_test_relationship("src/utils.rs", "src/lib.rs"),
-        ];
-
-        for rel in relationships {
-            matrix.add_relationship(rel);
-        }
-
-        let metrics = matrix.calculate_metrics();
-
-        assert_eq!(metrics.total_files, 4);
-        assert_eq!(metrics.total_relationships, 3);
-        assert_eq!(metrics.languages.len(), 2);
-        assert!(metrics.languages.contains(&"rust".to_string()));
-        assert!(metrics.languages.contains(&"python".to_string()));
-
-        // lib.rs should be the most highly coupled (3 incoming dependencies)
-        assert!(!metrics.highly_coupled_files.is_empty());
-        assert_eq!(
-            metrics.highly_coupled_files[0].0,
-            PathBuf::from("src/lib.rs")
-        );
-        assert_eq!(metrics.highly_coupled_files[0].1, 3); // 3 incoming edges
-    }
 }
