+++ tests/integration.rs
@@ -1,0 +1,1 @@
+mod rust;
