+++ tests/rust/cli/mod.rs
@@ -1,0 +1,5 @@
+// CLI module tests
+
+pub mod test_args;
+// Future CLI test modules would go here:
+// pub mod test_commands;
