+++ tests/rust/core/mod.rs
@@ -1,0 +1,7 @@
+// Core module tests
+
+pub mod test_matrix;
+pub mod test_scanner;
+
+// Future core test modules:
+// pub mod test_project;
