+++ tests/rust/mod.rs
@@ -1,0 +1,10 @@
+// Integration tests for CSD Rust code
+// This directory structure mirrors the src/ directory for organization
+
+pub mod cli;
+pub mod core;
+pub mod plugins;
+pub mod utils;
+
+// Common test utilities and helpers can go here
+pub mod common {}
