+++ tests/rust/plugins/mod.rs
@@ -1,0 +1,7 @@
+// Plugins module tests
+
+pub mod test_interface;
+
+// Future plugins test modules:
+// pub mod test_communication;
+// pub mod test_manager;
