+++ tests/rust/utils/mod.rs
@@ -1,0 +1,6 @@
+// Utils module tests
+
+pub mod test_config;
+
+// Future utils test modules:
+// pub mod test_file_utils;
